перед большими изменениями
