+++ js/main.js
@@ -1,5 +1,11 @@
 import './util.js';
 import './data.js';
-import './popup.js';
+import './similar-list.js';
 import './dialog.js';
 import './map.js';
+
+/*fetch('https://24.javascript.pages.academy/keksobooking/data')
+  .then((response) => response.json())
+  .then((notices) => {
+    console.log(notices);
+  });*/
